update CreationUtils to no longer need program.py
-Even more progress-
